Move `PlayerInfoForwarding` to the `api` module and remove duplicate `ServerInfoForwardingMode` in favor of `PlayerInfoForwarding`
