+++ api/src/main/java/com/velocitypowered/api/proxy/server/PlayerInfoForwarding.java
@@ -1,21 +1,11 @@
 /*
  * Copyright (C) 2018-2026 Velocity Contributors
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * The Velocity API is licensed under the terms of the MIT License. For more details,
+ * reference the LICENSE file in the api top-level directory.
  */
 
-package com.velocitypowered.proxy.config;
+package com.velocitypowered.api.proxy.server;
 
 /**
  * Supported player info forwarding methods.
